update dao and dto

+++ src/main/java/dao/CurrencyDao.java
@@ -5,6 +5,6 @@ import model.Currency;
 import java.util.Optional;
 
 public interface CurrencyDao extends CrudDao<Currency, String> {
-    Optional<Currency> findById(int id);
+    Optional<Currency> findById(Long id);
     Optional<Currency> findByCode(String code);
 }
